Removed NBNRequestKit version from podspec
In order for a project to use this Swift 2.3 branch, it also needs to use RequestKit’s Swift 2.3 branch by overriding the NBNRequestKit entry in the Podfile. This PR removes the version requirement from the NBNRequestKit dependency, resolving the conflict that arises when the Podfile specifies an external source for NBNRequestKit.