Make examples in README.md consistent
`session` is not declared in these examples and `task` result is not used. Also, these were not used consistently in all examples. In my opinion it would make more sense to clean those up, so that a beginner is not distracted from them, especially as `session` is an optional argument and `task` can be discarded.